Continued working towards implementing the entity system.
* Following tutorial at https://www.youtube.com/watch?v=Ool0zyg7iKg
* Added Engine engine to class MiniTownGame
* Created the initial shell for the EntityManager class.
* Updated Component class to say "public" for appropriate variables.
* Started work on PlayerInputSystem
* Started work on RenderSystem
* Created the initial PlayerCharacterFactory class.

I have started in ernist to get the intety system up and going.  As stated
above, I have started following through a tutorial on YouTube to get me off to
a solid start.  I am at ahout the 9:22 mark as I head to bed.  Obviously I am
not following the tutorial to the letter, but rather adapting it to my needs
as I go.  I have added a few TODO comments to the PlayerCharacterFactory that
pretty much summarize the minimum that I still need to do in order for the
player character (as it currently statnds) to be ready to push into the entity
system.

Once the input controlls are moved to the PlayerInputSystem and the PC and its
sprite are grouped together as an entity, the first real test will come.

-- burlingk@cagox.net